refactor: move recorder tests into module
