+++ src-tauri/src/audio_toolkit/audio/recorder.rs
@@ -1010,328 +1010,4 @@ fn run_consumer(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        is_microphone_access_denied, is_no_input_device_error, run_consumer, AudioRecorder,
-        CaptureTransportState, Cmd, VadPolicy,
-    };
-    use rtrb::RingBuffer;
-    use std::{
-        sync::{
-            atomic::{AtomicBool, Ordering},
-            mpsc, Arc, Mutex,
-        },
-        thread,
-        time::{Duration, Instant},
-    };
-
-    #[test]
-    fn unopened_recorder_does_not_need_reopen() {
-        let recorder = AudioRecorder::new().expect("recorder");
-        assert!(!recorder.needs_reopen());
-    }
-
-    #[test]
-    fn stream_error_requires_reopen() {
-        let recorder = AudioRecorder::new().expect("recorder");
-        recorder.stream_error.store(true, Ordering::Relaxed);
-        assert!(recorder.needs_reopen());
-    }
-
-    #[test]
-    fn shutdown_is_processed_without_audio_samples() {
-        let (_producer, consumer) = RingBuffer::<f32>::new(48_000);
-        let (cmd_tx, cmd_rx) = mpsc::channel();
-        let (done_tx, done_rx) = mpsc::channel();
-        let worker = thread::spawn(move || {
-            run_consumer(
-                48_000,
-                None,
-                consumer,
-                cmd_rx,
-                None,
-                None,
-                Arc::new(CaptureTransportState::default()),
-                Arc::new(AtomicBool::new(false)),
-                Instant::now(),
-            );
-            let _ = done_tx.send(());
-        });
-
-        cmd_tx.send(Cmd::Shutdown).expect("send shutdown");
-        assert!(done_rx.recv_timeout(Duration::from_secs(1)).is_ok());
-        worker.join().expect("join consumer");
-    }
-
-    #[test]
-    fn callback_writes_mono_samples() {
-        let (mut producer, mut consumer) = RingBuffer::<f32>::new(8);
-        let transport = CaptureTransportState::default();
-
-        AudioRecorder::write_input_to_ring(
-            &[0.25f32, -0.5, 1.0],
-            1,
-            None,
-            &mut producer,
-            &transport,
-        );
-
-        let mut output = [0.0; 3];
-        consumer.pop_entire_slice(&mut output).expect("samples");
-        assert_eq!(output, [0.25, -0.5, 1.0]);
-    }
-
-    #[test]
-    fn callback_downmixes_or_selects_multichannel_input() {
-        let transport = CaptureTransportState::default();
-        let (mut average_tx, mut average_rx) = RingBuffer::<f32>::new(4);
-        AudioRecorder::write_input_to_ring(
-            &[1.0f32, 3.0, -1.0, 1.0],
-            2,
-            None,
-            &mut average_tx,
-            &transport,
-        );
-        let mut averaged = [0.0; 2];
-        average_rx
-            .pop_entire_slice(&mut averaged)
-            .expect("averaged samples");
-        assert_eq!(averaged, [2.0, 0.0]);
-
-        let (mut selected_tx, mut selected_rx) = RingBuffer::<f32>::new(4);
-        AudioRecorder::write_input_to_ring(
-            &[1.0f32, 3.0, -1.0, 1.0],
-            2,
-            Some(1),
-            &mut selected_tx,
-            &transport,
-        );
-        let mut selected = [0.0; 2];
-        selected_rx
-            .pop_entire_slice(&mut selected)
-            .expect("selected samples");
-        assert_eq!(selected, [3.0, 1.0]);
-    }
-
-    #[test]
-    fn callback_acknowledges_pause_without_writing() {
-        let (mut producer, consumer) = RingBuffer::<f32>::new(4);
-        let transport = CaptureTransportState::default();
-        transport.pause_requested.store(true, Ordering::Release);
-
-        AudioRecorder::write_input_to_ring(&[1.0f32, 2.0], 1, None, &mut producer, &transport);
-
-        assert_eq!(consumer.slots(), 0);
-        assert!(transport.pause_acknowledged.load(Ordering::Acquire));
-    }
-
-    #[test]
-    fn post_write_pause_check_acknowledges_a_new_request() {
-        let transport = CaptureTransportState::default();
-        transport.pause_requested.store(true, Ordering::Release);
-        super::acknowledge_pause_after_write(&transport);
-        assert!(transport.pause_acknowledged.load(Ordering::Acquire));
-    }
-
-    #[test]
-    fn callback_partially_fills_ring_and_counts_dropped_audio() {
-        let (mut producer, mut consumer) = RingBuffer::<f32>::new(2);
-        let transport = CaptureTransportState::default();
-
-        AudioRecorder::write_input_to_ring(&[1.0f32, 2.0, 3.0], 1, None, &mut producer, &transport);
-
-        let mut captured = [0.0; 2];
-        consumer
-            .pop_entire_slice(&mut captured)
-            .expect("partial callback audio");
-        assert_eq!(captured, [1.0, 2.0]);
-        assert_eq!(transport.overrun_samples.load(Ordering::Relaxed), 1);
-    }
-
-    #[test]
-    fn bounded_drain_leaves_remaining_samples_for_the_next_command_cycle() {
-        let (mut producer, mut consumer) = RingBuffer::<f32>::new(8);
-        producer
-            .push_entire_slice(&[1.0, 2.0, 3.0, 4.0, 5.0])
-            .expect("samples");
-        let mut drained = Vec::new();
-
-        let count = super::drain_available_samples(&mut consumer, 3, |part| {
-            drained.extend_from_slice(part)
-        });
-
-        assert_eq!(count, 3);
-        assert_eq!(drained, [1.0, 2.0, 3.0]);
-        assert_eq!(consumer.slots(), 2);
-    }
-
-    #[test]
-    fn ring_wraparound_preserves_both_read_slices_in_order() {
-        let (mut producer, mut consumer) = RingBuffer::<f32>::new(5);
-        let transport = CaptureTransportState::default();
-        producer
-            .push_entire_slice(&[1.0, 2.0, 3.0, 4.0])
-            .expect("initial samples");
-        let mut discarded = [0.0; 3];
-        consumer
-            .pop_entire_slice(&mut discarded)
-            .expect("advance ring head");
-
-        AudioRecorder::write_input_to_ring(
-            &[5.0f32, 6.0, 7.0, 8.0],
-            1,
-            None,
-            &mut producer,
-            &transport,
-        );
-
-        let chunk = consumer.read_chunk(5).expect("wrapped samples");
-        let (first, second) = chunk.as_slices();
-        assert!(!first.is_empty());
-        assert!(!second.is_empty());
-        let ordered = first
-            .iter()
-            .chain(second.iter())
-            .copied()
-            .collect::<Vec<_>>();
-        assert_eq!(ordered, [4.0, 5.0, 6.0, 7.0, 8.0]);
-    }
-
-    #[test]
-    fn stop_drains_samples_before_replying_and_streaming_sees_the_same_order() {
-        let (mut producer, consumer) = RingBuffer::<f32>::new(16_000);
-        let transport = Arc::new(CaptureTransportState::default());
-        let (cmd_tx, cmd_rx) = mpsc::channel();
-        let streamed = Arc::new(Mutex::new(Vec::new()));
-        let streamed_cb = Arc::clone(&streamed);
-        let consumer_transport = Arc::clone(&transport);
-        let worker = thread::spawn(move || {
-            run_consumer(
-                16_000,
-                None,
-                consumer,
-                cmd_rx,
-                None,
-                Some(Arc::new(move |frame| {
-                    streamed_cb.lock().unwrap().extend_from_slice(frame)
-                })),
-                consumer_transport,
-                Arc::new(AtomicBool::new(false)),
-                Instant::now(),
-            );
-        });
-
-        let (ready_tx, ready_rx) = mpsc::channel();
-        cmd_tx
-            .send(Cmd::Start(VadPolicy::Disabled, Instant::now(), ready_tx))
-            .expect("start");
-        AudioRecorder::write_input_to_ring(
-            &[0.25f32, -0.5, 1.0],
-            1,
-            None,
-            &mut producer,
-            &transport,
-        );
-        ready_rx
-            .recv_timeout(Duration::from_secs(1))
-            .expect("capture ready");
-
-        let (reply_tx, reply_rx) = mpsc::channel();
-        cmd_tx.send(Cmd::Stop(reply_tx)).expect("stop");
-        let deadline = Instant::now() + Duration::from_secs(1);
-        while !transport.pause_requested.load(Ordering::Acquire) {
-            assert!(Instant::now() < deadline, "pause was not requested");
-            thread::sleep(Duration::from_millis(1));
-        }
-        // The first callback after Stop acknowledges the boundary and must not
-        // append samples behind it.
-        AudioRecorder::write_input_to_ring(&[99.0f32], 1, None, &mut producer, &transport);
-
-        let samples = reply_rx
-            .recv_timeout(Duration::from_secs(1))
-            .expect("stop reply");
-        assert_eq!(&samples[..3], &[0.25, -0.5, 1.0]);
-        assert!(!samples.contains(&99.0));
-        let streamed = streamed.lock().unwrap();
-        assert_eq!(&streamed[..3], &[0.25, -0.5, 1.0]);
-        assert!(!streamed.contains(&99.0));
-
-        cmd_tx.send(Cmd::Shutdown).expect("shutdown");
-        worker.join().expect("consumer worker");
-    }
-
-    #[test]
-    fn missing_callback_at_stop_marks_stream_for_rebuild_and_returns_samples() {
-        let (_producer, consumer) = RingBuffer::<f32>::new(16_000);
-        let transport = Arc::new(CaptureTransportState::default());
-        let stream_error = Arc::new(AtomicBool::new(false));
-        let observed_error = Arc::clone(&stream_error);
-        let (cmd_tx, cmd_rx) = mpsc::channel();
-        let worker_transport = Arc::clone(&transport);
-        let worker = thread::spawn(move || {
-            run_consumer(
-                16_000,
-                None,
-                consumer,
-                cmd_rx,
-                None,
-                None,
-                worker_transport,
-                stream_error,
-                Instant::now(),
-            );
-        });
-
-        let (ready_tx, _ready_rx) = mpsc::channel();
-        cmd_tx
-            .send(Cmd::Start(VadPolicy::Disabled, Instant::now(), ready_tx))
-            .expect("start");
-        let (reply_tx, reply_rx) = mpsc::channel();
-        cmd_tx.send(Cmd::Stop(reply_tx)).expect("stop");
-
-        let samples = reply_rx
-            .recv_timeout(Duration::from_secs(3))
-            .expect("pause timeout still returns captured samples");
-        assert!(samples.is_empty());
-        worker.join().expect("consumer exits after pause timeout");
-        assert!(observed_error.load(Ordering::Acquire));
-    }
-
-    #[test]
-    fn detects_access_is_denied() {
-        assert!(is_microphone_access_denied("Access is denied"));
-    }
-
-    #[test]
-    fn detects_permission_denied() {
-        assert!(is_microphone_access_denied("permission denied"));
-    }
-
-    #[test]
-    fn detects_windows_error_code() {
-        assert!(is_microphone_access_denied("WASAPI error: 0x80070005"));
-    }
-
-    #[test]
-    fn does_not_match_unrelated_errors() {
-        assert!(!is_microphone_access_denied("device not found"));
-    }
-
-    #[test]
-    fn detects_no_input_device() {
-        assert!(is_no_input_device_error("No input device found"));
-    }
-
-    #[test]
-    fn detects_coreaudio_config_error() {
-        assert!(is_no_input_device_error(
-            "Failed to fetch preferred config: A backend-specific error has occurred: An unknown error unknown to the coreaudio-rs API occurred"
-        ));
-    }
-
-    #[test]
-    fn does_not_match_other_errors_for_no_device() {
-        assert!(!is_no_input_device_error("permission denied"));
-        assert!(!is_no_input_device_error("device not found"));
-    }
-}
+mod tests;
